+++ src/ra_proc_macro.rs
@@ -1,103 +1,18 @@
-mod tt;
-
-use crate::{process::ProcessBuilderExt as _, shell::Shell};
 use anyhow::{anyhow, bail, Context as _};
-use camino::{Utf8Path, Utf8PathBuf};
 use cargo_metadata as cm;
-use cargo_util::ProcessBuilder;
+use itertools::chain;
 use maplit::btreemap;
-use serde::{Deserialize, Serialize};
-use std::{
-    collections::{BTreeMap, BTreeSet, VecDeque},
-    env,
-    io::{self, BufRead as _, BufReader, Read as _, Write as _},
-    path::{Path, PathBuf},
-    process::{Child, ChildStdin, ChildStdout, Command, Stdio},
+use ra_ap_paths::AbsPath;
+use ra_ap_proc_macro_api::{
+    msg::PanicMessage, MacroDylib, ProcMacro, ProcMacroKind, ProcMacroServer,
 };
+use ra_ap_tt::{self as tt, DelimiterKind, Leaf, TokenId};
+use std::collections::{BTreeMap, BTreeSet};
 
-pub(crate) fn dl_ra(
-    dir: &Path,
-    rustc_version: &cm::Version,
-    shell: &mut Shell,
-) -> anyhow::Result<PathBuf> {
-    cargo_util::paths::create_dir_all(dir)?;
-
-    let tag = if rustc_version.to_string() == "1.47.0" {
-        &TAG_FOR_EQ_1_47
-    } else {
-        &TAG_FOR_GEQ_1_48
-    };
-
-    let ra_path = dir
-        .join(format!("rust-analyzer-{}", tag))
-        .with_extension(env::consts::EXE_EXTENSION);
-
-    if ra_path.exists() {
-        return Ok(ra_path);
-    }
-
-    let file_name = if cfg!(all(target_arch = "x86_64", target_os = "windows")) {
-        "rust-analyzer-x86_64-pc-windows-msvc.gz"
-    } else if cfg!(all(target_arch = "x86_64", target_os = "macos")) {
-        "rust-analyzer-x86_64-apple-darwin.gz"
-    } else if cfg!(all(target_arch = "x86_64", target_os = "linux")) {
-        "rust-analyzer-x86_64-unknown-linux-gnu.gz"
-    } else if cfg!(all(target_arch = "aarch64", target_os = "windows")) {
-        "rust-analyzer-aarch64-pc-windows-msvc.gz"
-    } else if cfg!(all(target_arch = "aarch64", target_os = "macos")) {
-        "rust-analyzer-aarch64-apple-darwin.gz"
-    } else if cfg!(all(target_arch = "aarch64", target_os = "linux")) {
-        "rust-analyzer-aarch64-unknown-linux-gnu.gz"
-    } else {
-        bail!("only x86_64/aarch64 Windows/macOS/Linux supported");
-    };
-
-    let url = format!(
-        "https://github.com/rust-analyzer/rust-analyzer/releases/download/{}/{}",
-        tag, file_name,
-    );
-    let gz = &curl(&url, dir, shell)?;
-    let ra = decode_gz(gz).with_context(|| format!("could not decode {}", file_name))?;
-    cargo_util::paths::write(&ra_path, ra)?;
-    shell.status("Wrote", ra_path.display())?;
-    chmod755(&ra_path)?;
-    return Ok(ra_path);
-
-    static TAG_FOR_EQ_1_47: &str = "2021-07-12";
-    static TAG_FOR_GEQ_1_48: &str = "2021-08-09";
-
-    fn curl(url: &str, cwd: &Path, shell: &mut Shell) -> anyhow::Result<Vec<u8>> {
-        let curl_exe = which::which("curl").map_err(|_| anyhow!("command not found: curl"))?;
-        ProcessBuilder::new(curl_exe)
-            .args(&[url, "-L"])
-            .cwd(cwd)
-            .try_inspect(|this| shell.status("Running", this))?
-            .read_stdout()
-    }
-
-    fn decode_gz(data: &[u8]) -> io::Result<Vec<u8>> {
-        let mut buf = vec![];
-        libflate::gzip::Decoder::new(data)?.read_to_end(&mut buf)?;
-        Ok(buf)
-    }
-
-    #[cfg(unix)]
-    fn chmod755(path: &Path) -> io::Result<()> {
-        use std::{fs::Permissions, os::unix::fs::PermissionsExt as _};
-        std::fs::set_permissions(path, Permissions::from_mode(0o755))
-    }
-
-    #[cfg(not(unix))]
-    #[allow(clippy::unnecessary_wraps)]
-    fn chmod755(_: &Path) -> io::Result<()> {
-        Ok(())
-    }
-}
-
-pub(crate) fn list_proc_macro_dlls<P: FnMut(&cm::PackageId) -> bool>(
+pub(crate) fn list_proc_macro_dylibs<P: FnMut(&cm::PackageId) -> bool>(
     cargo_messages: &[cm::Message],
     mut filter: P,
-) -> BTreeMap<&cm::PackageId, &Utf8Path> {
+) -> BTreeMap<&cm::PackageId, &AbsPath> {
     cargo_messages
         .iter()
         .flat_map(|message| match message {
@@ -111,321 +26,232 @@ pub(crate) fn list_proc_macro_dlls<P: FnMut(&cm::PackageId) -> bool>(
                  package_id,
                  filenames,
                  ..
-             }| filenames.get(0).map(|filename| (package_id, &**filename)),
+             }| {
+                filenames
+                    .get(0)
+                    .map(|filename| (package_id, AbsPath::assert(filename.as_ref())))
+            },
         )
         .collect()
 }
 
-pub(crate) struct ProcMacroExpander<'msg> {
-    ra: RaProcMacro,
-    func_like: BTreeMap<String, (&'msg cm::PackageId, &'msg Utf8Path)>,
-    attr: BTreeMap<String, (&'msg cm::PackageId, &'msg Utf8Path)>,
-    custom_derive: BTreeMap<String, (&'msg cm::PackageId, &'msg Utf8Path)>,
+pub struct ProcMacroExpander<'msg> {
+    custom_derive: BTreeMap<String, (&'msg cm::PackageId, ProcMacro)>,
+    func_like: BTreeMap<String, (&'msg cm::PackageId, ProcMacro)>,
+    attr: BTreeMap<String, (&'msg cm::PackageId, ProcMacro)>,
 }
 
 impl<'msg> ProcMacroExpander<'msg> {
-    pub(crate) fn new(
-        rust_analyzer_exe: &Path,
-        dll_paths: &BTreeMap<&'msg cm::PackageId, &'msg Utf8Path>,
-        shell: &mut Shell,
+    pub(crate) fn spawn(
+        cargo_equip_exe: &AbsPath,
+        dylib_paths: &BTreeMap<&'msg cm::PackageId, &'msg AbsPath>,
     ) -> anyhow::Result<Self> {
-        shell.status(
-            "Spawning",
-            format!("`{} proc-macro`", rust_analyzer_exe.display()),
-        )?;
+        let server =
+            ProcMacroServer::spawn(cargo_equip_exe.to_path_buf(), ["rust-analyzer-proc-macro"])?;
 
-        let mut this = Self {
-            ra: RaProcMacro::new(rust_analyzer_exe)?,
-            func_like: btreemap!(),
-            attr: btreemap!(),
-            custom_derive: btreemap!(),
-        };
+        let mut custom_derive = btreemap!();
+        let mut func_like = btreemap!();
+        let mut attr = btreemap!();
 
-        for (package_id, dll) in dll_paths {
-            for (name, kind) in this.ra.list_macro(dll, |msg| {
-                shell.warn(format!("error from RA: {}", msg))?;
-                Ok(())
-            })? {
-                match kind {
-                    ProcMacroKind::CustomDerive => {
-                        if this
-                            .custom_derive
-                            .insert(name.clone(), (*package_id, *dll))
-                            .is_some()
-                        {
-                            bail!("duplicated `#[derive({})]`", name);
-                        }
-                    }
-                    ProcMacroKind::FuncLike => {
-                        if this
-                            .func_like
-                            .insert(name.clone(), (*package_id, *dll))
-                            .is_some()
-                        {
-                            bail!("duplicated `{}!`", name);
-                        }
-                    }
-                    ProcMacroKind::Attr => {
-                        if this
-                            .attr
-                            .insert(name.clone(), (*package_id, *dll))
-                            .is_some()
-                        {
-                            bail!("duplicated `#[{}]`", name);
-                        }
-                    }
+        for (&package_id, dylib_path) in dylib_paths {
+            let result = MacroDylib::new(dylib_path.to_path_buf())
+                .map_err(anyhow::Error::from)
+                .and_then(|d| server.load_dylib(d).map_err(|e| anyhow!("{}", e)))
+                .with_context(|| "rust-analyzer error")?;
+            let proc_macros = match result {
+                Ok(proc_macros) => proc_macros,
+                Err(err) => bail!("{}", err),
+            };
+            for proc_macro in proc_macros {
+                match proc_macro.kind() {
+                    ProcMacroKind::CustomDerive => &mut custom_derive,
+                    ProcMacroKind::FuncLike => &mut func_like,
+                    ProcMacroKind::Attr => &mut attr,
                 }
+                .insert(proc_macro.name().to_owned(), (package_id, proc_macro));
             }
         }
 
-        for name in this.func_like.keys() {
-            shell.status("Readied", format!("`{}!`", name))?;
-        }
-        for name in this.attr.keys() {
-            shell.status("Readied", format!("`#[{}]`", name))?;
-        }
-        for name in this.custom_derive.keys() {
-            shell.status("Readied", format!("`#[derive({})]`", name))?;
-        }
-
-        Ok(this)
+        Ok(Self {
+            custom_derive,
+            func_like,
+            attr,
+        })
     }
 
     pub(crate) fn macro_names(
         &self,
     ) -> impl Iterator<Item = (&'msg cm::PackageId, BTreeSet<&str>)> {
         let mut names = BTreeMap::<_, BTreeSet<_>>::new();
-        for (name, (pkg, _)) in self
-            .func_like
-            .iter()
-            .chain(&self.attr)
-            .chain(&self.custom_derive)
-        {
-            names.entry(*pkg).or_default().insert(&**name);
+        for (name, &(pkg, _)) in chain!(&self.custom_derive, &self.func_like, &self.attr) {
+            names.entry(pkg).or_default().insert(&**name);
         }
         names.into_iter()
     }
 
-    pub(crate) fn expand_func_like_macro(
+    pub(crate) fn attempt_expand_custom_derive(
         &mut self,
         name: &str,
         body: impl FnOnce() -> proc_macro2::TokenStream,
-        on_error: impl FnMut(&str) -> anyhow::Result<()>,
     ) -> anyhow::Result<Option<proc_macro2::Group>> {
-        if let Some((_, dll)) = self.func_like.get(name) {
-            let dll = *dll;
-            self.expand(dll, name, body(), None, on_error).map(Some)
-        } else {
-            Ok(None)
-        }
+        self.attempt_expand(name, ProcMacroKind::CustomDerive, body, None::<fn() -> _>)
     }
 
-    pub(crate) fn expand_attr_macro(
+    pub(crate) fn attempt_expand_func_like(
+        &mut self,
+        name: &str,
+        body: impl FnOnce() -> proc_macro2::TokenStream,
+    ) -> anyhow::Result<Option<proc_macro2::Group>> {
+        self.attempt_expand(name, ProcMacroKind::FuncLike, body, None::<fn() -> _>)
+    }
+
+    pub(crate) fn attempt_expand_attr(
         &mut self,
         name: &str,
         body: impl FnOnce() -> proc_macro2::TokenStream,
         attr: impl FnOnce() -> proc_macro2::Group,
-        on_error: impl FnMut(&str) -> anyhow::Result<()>,
     ) -> anyhow::Result<Option<proc_macro2::Group>> {
-        if let Some((_, dll)) = self.attr.get(name) {
-            let dll = *dll;
-            self.expand(dll, name, body(), Some(attr()), on_error)
-                .map(Some)
-        } else {
-            Ok(None)
-        }
+        self.attempt_expand(name, ProcMacroKind::Attr, body, Some(attr))
     }
 
-    pub(crate) fn expand_derive_macro(
-        &mut self,
+    fn attempt_expand(
+        &self,
         name: &str,
-        body: impl FnOnce() -> proc_macro2::TokenStream,
-        on_error: impl FnMut(&str) -> anyhow::Result<()>,
+        kind: ProcMacroKind,
+        subtree: impl FnOnce() -> proc_macro2::TokenStream,
+        attr: Option<impl FnOnce() -> proc_macro2::Group>,
     ) -> anyhow::Result<Option<proc_macro2::Group>> {
-        if let Some((_, dll)) = self.custom_derive.get(name) {
-            let dll = *dll;
-            self.expand(dll, name, body(), None, on_error).map(Some)
-        } else {
-            Ok(None)
+        match kind {
+            ProcMacroKind::CustomDerive => &self.custom_derive,
+            ProcMacroKind::FuncLike => &self.func_like,
+            ProcMacroKind::Attr => &self.attr,
         }
-    }
-
-    fn expand(
-        &mut self,
-        dll_path: &Utf8Path,
-        macro_name: &str,
-        macro_body: proc_macro2::TokenStream,
-        attributes: Option<proc_macro2::Group>,
-        on_error: impl FnMut(&str) -> anyhow::Result<()>,
-    ) -> anyhow::Result<proc_macro2::Group> {
-        self.ra
-            .expansion_macro(
-                dll_path,
-                macro_name,
-                proc_macro2::Group::new(proc_macro2::Delimiter::None, macro_body).into(),
-                attributes.map(Into::into),
-                on_error,
-            )
-            .map(Into::into)
-    }
-}
-
-pub(crate) struct RaProcMacro {
-    process_status: Child,
-    process_stdin: ChildStdin,
-    process_stdout: BufReader<ChildStdout>,
-    list_macro_responses: VecDeque<ListMacrosResult>,
-    expansion_macro_responses: VecDeque<ExpansionResult>,
-}
-
-impl RaProcMacro {
-    pub(crate) fn new(rust_analyzer_exe: &Path) -> anyhow::Result<Self> {
-        let mut process = Command::new(rust_analyzer_exe)
-            .arg("proc-macro")
-            .stdin(Stdio::piped())
-            .stdout(Stdio::piped())
-            .spawn()
-            .with_context(|| format!("could not execute `{}`", rust_analyzer_exe.display()))?;
-
-        let process_stdin = process.stdin.take().expect("specified `Stdio::piped()`");
-        let process_stdout = process.stdout.take().expect("specified `Stdio::piped()`");
-
-        Ok(Self {
-            process_status: process,
-            process_stdin,
-            process_stdout: BufReader::new(process_stdout),
-            list_macro_responses: VecDeque::new(),
-            expansion_macro_responses: VecDeque::new(),
+        .get(name)
+        .map(|(_, proc_macro)| {
+            let output = &proc_macro
+                .expand(
+                    &from_proc_macro2_group(&proc_macro2::Group::new(
+                        proc_macro2::Delimiter::None,
+                        subtree(),
+                    )),
+                    attr.map(|f| from_proc_macro2_group(&f())).as_ref(),
+                    vec![],
+                )
+                .map_err(|e| anyhow!("{}", e))
+                .with_context(|| "rust-analyzer error")?
+                .map_err(|PanicMessage(s)| anyhow!("proc macro paniced: {s:?}"))?;
+            Ok(from_ra_subtree(output))
         })
+        .transpose()
+    }
+}
+
+fn from_proc_macro2_group(group: &proc_macro2::Group) -> tt::Subtree {
+    return tt::Subtree {
+        delimiter: from_proc_macro2_delimiter(group.delimiter()),
+        token_trees: group
+            .stream()
+            .into_iter()
+            .map(|tt| from_proc_macro2_token_tree(&tt))
+            .collect(),
+    };
+
+    fn from_proc_macro2_delimiter(delimiter: proc_macro2::Delimiter) -> Option<tt::Delimiter> {
+        return match delimiter {
+            proc_macro2::Delimiter::Parenthesis => Some(from(DelimiterKind::Parenthesis)),
+            proc_macro2::Delimiter::Brace => Some(from(DelimiterKind::Brace)),
+            proc_macro2::Delimiter::Bracket => Some(from(DelimiterKind::Bracket)),
+            proc_macro2::Delimiter::None => None,
+        };
+
+        fn from(kind: DelimiterKind) -> tt::Delimiter {
+            tt::Delimiter {
+                id: TokenId::unspecified(),
+                kind,
+            }
+        }
     }
 
-    pub(crate) fn list_macro(
-        &mut self,
-        dll_path: &Utf8Path,
-        mut on_error: impl FnMut(&str) -> anyhow::Result<()>,
-    ) -> anyhow::Result<Vec<(String, ProcMacroKind)>> {
-        self.request(Request::ListMacro(ListMacrosTask {
-            lib: dll_path.to_owned(),
-        }))?;
-
-        loop {
-            if let Some(ListMacrosResult { macros }) = self.list_macro_responses.pop_front() {
-                break Ok(macros);
-            }
-            self.wait_response(&mut on_error)?;
+    fn from_proc_macro2_token_tree(tt: &proc_macro2::TokenTree) -> tt::TokenTree {
+        match tt {
+            proc_macro2::TokenTree::Group(g) => from_proc_macro2_group(g).into(),
+            proc_macro2::TokenTree::Ident(i) => Leaf::from(from_proc_macro2_ident(i)).into(),
+            proc_macro2::TokenTree::Punct(p) => Leaf::from(from_proc_macro2_punct(p)).into(),
+            proc_macro2::TokenTree::Literal(l) => Leaf::from(from_proc_macro2_literal(l)).into(),
         }
     }
 
-    fn expansion_macro(
-        &mut self,
-        dll_path: &Utf8Path,
-        macro_name: &str,
-        macro_body: tt::Subtree,
-        attributes: Option<tt::Subtree>,
-        mut on_error: impl FnMut(&str) -> anyhow::Result<()>,
-    ) -> anyhow::Result<tt::Subtree> {
-        self.request(Request::ExpansionMacro(ExpansionTask {
-            macro_body,
-            macro_name: macro_name.to_owned(),
-            attributes,
-            lib: dll_path.to_owned(),
-            env: vec![],
-        }))?;
-
-        loop {
-            if let Some(ExpansionResult { expansion }) = self.expansion_macro_responses.pop_front()
-            {
-                break Ok(expansion);
-            }
-            self.wait_response(&mut on_error)?;
+    fn from_proc_macro2_ident(ident: &proc_macro2::Ident) -> tt::Ident {
+        tt::Ident {
+            text: ident.to_string().into(),
+            id: TokenId::unspecified(),
         }
     }
 
-    fn request(&mut self, req: Request) -> io::Result<()> {
-        let mut req = serde_json::to_string(&req).expect("should not fail");
-        req += "\n";
-        self.process_stdin.write_all(req.as_ref())?;
-        self.process_stdin.flush()
+    fn from_proc_macro2_punct(punct: &proc_macro2::Punct) -> tt::Punct {
+        tt::Punct {
+            char: punct.as_char(),
+            spacing: from_proc_macro2_spacing(punct.spacing()),
+            id: TokenId::unspecified(),
+        }
     }
 
-    fn wait_response(
-        &mut self,
-        mut on_error: impl FnMut(&str) -> anyhow::Result<()>,
-    ) -> anyhow::Result<()> {
-        let res = &mut "".to_owned();
-        self.process_stdout.read_line(res)?;
-        if res.is_empty() {
-            if let Some(status) = self.process_status.try_wait()? {
-                bail!("rust-analyzer unexpectedly terminated ({})", status);
-            }
+    fn from_proc_macro2_spacing(spacing: proc_macro2::Spacing) -> tt::Spacing {
+        match spacing {
+            proc_macro2::Spacing::Alone => tt::Spacing::Alone,
+            proc_macro2::Spacing::Joint => tt::Spacing::Joint,
         }
-        match serde_json::from_str(res)
-            .with_context(|| "could not deserialize values from rust-analyzer")?
-        {
-            Response::Error(err) => {
-                on_error(&serde_json::to_string(&err).expect("should not fail"))?
-            }
-            Response::ListMacro(res) => self.list_macro_responses.push_back(res),
-            Response::ExpansionMacro(res) => self.expansion_macro_responses.push_back(res),
+    }
+
+    fn from_proc_macro2_literal(lit: &proc_macro2::Literal) -> tt::Literal {
+        tt::Literal {
+            text: lit.to_string().into(),
+            id: TokenId::unspecified(),
         }
-        Ok(())
     }
 }
 
-// https://github.com/rust-analyzer/rust-analyzer/blob/2021-03-29/crates/proc_macro_api/src/msg.rs
+fn from_ra_subtree(subtree: &tt::Subtree) -> proc_macro2::Group {
+    return proc_macro2::Group::new(
+        from_ra_delimiter(subtree.delimiter),
+        subtree.token_trees.iter().map(from_ra_token_tree).collect(),
+    );
 
-#[derive(Serialize)]
-enum Request {
-    ListMacro(ListMacrosTask),
-    ExpansionMacro(ExpansionTask),
-}
+    fn from_ra_delimiter(delimiter: Option<tt::Delimiter>) -> proc_macro2::Delimiter {
+        match delimiter.map(|tt::Delimiter { kind, .. }| kind) {
+            Some(DelimiterKind::Parenthesis) => proc_macro2::Delimiter::Parenthesis,
+            Some(DelimiterKind::Brace) => proc_macro2::Delimiter::Brace,
+            Some(DelimiterKind::Bracket) => proc_macro2::Delimiter::Bracket,
+            None => proc_macro2::Delimiter::None,
+        }
+    }
 
-#[derive(Serialize)]
-struct ListMacrosTask {
-    lib: Utf8PathBuf,
-}
+    fn from_ra_token_tree(tt: &tt::TokenTree) -> proc_macro2::TokenTree {
+        match tt {
+            tt::TokenTree::Subtree(s) => proc_macro2::TokenTree::Group(from_ra_subtree(s)),
+            tt::TokenTree::Leaf(Leaf::Ident(i)) => from_ra_ident(i).into(),
+            &tt::TokenTree::Leaf(Leaf::Punct(p)) => from_ra_punct(p).into(),
+            tt::TokenTree::Leaf(Leaf::Literal(l)) => from_ra_literal(l).into(),
+        }
+    }
 
-#[derive(Serialize)]
-struct ExpansionTask {
-    macro_body: tt::Subtree,
-    macro_name: String,
-    attributes: Option<tt::Subtree>,
-    lib: Utf8PathBuf,
-    env: Vec<(String, String)>,
-}
+    fn from_ra_ident(ident: &tt::Ident) -> proc_macro2::Ident {
+        proc_macro2::Ident::new(&ident.text, proc_macro2::Span::call_site())
+    }
 
-#[derive(Deserialize)]
-enum Response {
-    Error(ResponseError),
-    ListMacro(ListMacrosResult),
-    ExpansionMacro(ExpansionResult),
-}
+    fn from_ra_punct(punct: tt::Punct) -> proc_macro2::Punct {
+        proc_macro2::Punct::new(punct.char, from_ra_spacing(punct.spacing))
+    }
 
-#[derive(Deserialize, Serialize)]
-struct ResponseError {
-    code: ErrorCode,
-    message: String,
-}
+    fn from_ra_spacing(spacing: tt::Spacing) -> proc_macro2::Spacing {
+        match spacing {
+            tt::Spacing::Alone => proc_macro2::Spacing::Alone,
+            tt::Spacing::Joint => proc_macro2::Spacing::Joint,
+        }
+    }
 
-#[derive(Copy, Clone, Deserialize, Serialize)]
-enum ErrorCode {
-    ServerErrorEnd,
-    ExpansionError,
-}
-
-#[derive(Deserialize)]
-struct ListMacrosResult {
-    macros: Vec<(String, ProcMacroKind)>,
-}
-
-#[derive(Copy, Clone, Deserialize, Debug)]
-pub(crate) enum ProcMacroKind {
-    CustomDerive,
-    FuncLike,
-    Attr,
-}
-
-#[derive(Deserialize)]
-struct ExpansionResult {
-    expansion: tt::Subtree,
+    fn from_ra_literal(lit: &tt::Literal) -> proc_macro2::Literal {
+        syn::parse_str(&lit.text)
+            .unwrap_or_else(|e| panic!("could not parse {:?} as a literal: {}", &lit.text, e))
+    }
 }